+++ sdk/cognitivelanguage/azure-ai-language-conversations/tests/async/test_conv_summarization_async.py
@@ -8,20 +8,20 @@ import pytest
 
 from azure.core.exceptions import HttpResponseError, ClientAuthenticationError
 from azure.core.credentials import AzureKeyCredential
-from testcase import (
-    ConversationTest,
-    GlobalConversationAccountPreparer
-)
-from azure.ai.language.conversations import ConversationAnalysisClient
 
-class ConversationalIssueSummarizationTests(ConversationTest):
+from testcase import GlobalConversationAccountPreparer
+from asynctestcase import AsyncConversationTest
+from azure.ai.language.conversations.aio import ConversationAnalysisClient
+
+
+class ConversationalSummarizationAsyncTests(AsyncConversationTest):
 
     @GlobalConversationAccountPreparer()
-    def test_conversational_issue_summarization(self, endpoint, key):
+    async def test_conversational_summarization(self, endpoint, key):
         # analyze query
         client = ConversationAnalysisClient(endpoint, AzureKeyCredential(key))
-        with client:
-            poller = client.begin_conversation_analysis(
+        async with client:
+            poller = await client.begin_conversation_analysis(
                 task={
                     "displayName": "Analyze conversations from xxx",
                     "analysisInput": {
@@ -66,7 +66,7 @@ class ConversationalIssueSummarizationTests(ConversationTest):
             )
         
             # assert - main object
-            result = poller.result()
+            result = await poller.result()
             assert not result is None
             assert result["status"] == "succeeded"